At least it works now on WOW64 (after DISABLING manual-map)
/ Fixed x86 build compilation errors
/ Fixed NTSTATUS type from ULONG to UINT

*** RtlQueryProcessHeapInformation and RtlQueryProcessDebugInformation works perfectly fine on x86, but not on x64.
I think the struct DEBUG_BUFFER and RTL_DEBUG_INFORMATION has wrong types.
Ill try to dump those structs again by myself (windbg !dt)

+++ AntiDebugLib/Native/NativeDefs+NTSTATUS.cs
@@ -4,7 +4,7 @@
     /// https://learn.microsoft.com/en-us/openspecs/windows_protocols/ms-erref/596a1078-e883-4972-9bbc-49e60bebca55
     /// Used Copytables 0.5.9
     /// </summary>
-    public enum NTSTATUS : ulong
+    public enum NTSTATUS : uint
     {
         STATUS_SUCCESS = 0x00000000,
         STATUS_WAIT_0 = 0x00000000,
